Add collapsable navbar on screens up to Medium
The previous "Header" had a layout configured by Row and Col components.
Created the "NavbarComponent" using actual navbar components from
reactsrap instead. This made setting it up collapsable links with a
hamburger button easier, and the code looks cleaner as well.

Additionally, minor changes to the write form and validation and Admin
page was made for cleaner code and functionality.