Prep express back end for Google Cloud Function deployment
Google Cloud Function injects port 8080 as the .env port as part of
their build process. Somewhere along the way this was causing a strange
error saying that port 8080 was already in use.

The port ultimately seems irrelevant. After commenting out the part of
www.js setting the port as the ".env" port and setting it to just use
3001, calls to the back end at the google address work without
specifying a port at all.

Front end was also changed to use the google cloud function.