Display author name/date
Added a route to express and utility operation on the front end to
pull the author's first and last name and username out of the databse.
I created a new route entirely rather than use the existing one becuase
I want a route to exist to get ALL user information that is gated
behind checking if the user is authorized. To just get the name and
username should be public info though, hence the separate route.