Track comments on Article.
Each comment "document" lists the articleId it's matched with.
However, it also makes sense that articles would track a list of
comments. Otherwise, every time you need a list of comments for an
article, you would have to search every single comment and check each
and every articleId. This way the most common use--finding comments
from a specific article as a reference--gives us only the specific
comments that apply.