Comments pseudo-lazy-load
The entire list of an articles comments is currently sent to the user.
Then only 4 comments at a time are shown to prevent the page from
getting super long if there are a lot of comments.

The user can optionally load more comments by clicking the button
at the bottom, which redefines "displayedComments", and updates the
"page" for the next time it updates.

The button to load comments will stop showing when the length of
displayedComments is the same (or more than as a fail safe) as comments.

Added an "Icing on the cake" task of lazy-loading on the back end.
To see why this might make a difference, check the logs of the server
when a page loads. You will see that a "/users/[user _id]/publiclookup
is done for EVERY comment, even though only a partial list are shown.