Populate comment author
Route to get specific comment now pu=opulates the author.

Previously the front end was making multiple requests. It would make an
api request to get the comment. Then it would make a new request to
for EACH of the comments to get the "public info" for each user.

Now, the back end just populates the author info on the initial call.