Allow Write page to be used for editing
Added props for "articleBeingEdited" to be passed.

If "articleBeingEdited" is undefined, it will
post a new article.

If the prop, "articleBeingEdited" IS passed,
this page will instead use the values of the
article being passed in, and use a "PUT" HTTP
method instead of "POST".

The Url suffix will also be changed to
"/articles/[articleBeingEdited._id]"