Users who are premium cannot reach joinup page
The JoinUp page checks the current user.

If the current user is undefined, it means they
are not signed in, and it will request they do so.

If the current user is not a premium user, it will
display a message and paypal/venmo links to
complete the transaction.

If the user is already a premium user, it will
display a message informing them of that fact.

The rest of the application should not allow a
premium user to find any links to the /joinup page,
but this is added protection against accidentally
arriving at this page manually or by using "back.""